# swagger.yaml
openapi: 3.0.3
info:
  title: Nexus Project Management API
  description: |
    RESTful API for the Nexus Project Management System.
    
    ## Authentication
    This API uses JWT (JSON Web Token) for authentication. After registering or logging in, 
    you will receive a JWT token that must be included in the `Authorization` header for all 
    protected endpoints.
    
    ## Authorization
    The API implements role-based access control (RBAC) with two roles:
    - **ADMIN**: Full access to all resources
    - **USER**: Limited access to assigned resources only
    
    ## Features
    - User authentication and management
    - Project management with team assignments
    - Task management with status tracking
    - Role-based authorization
    - Pagination and search capabilities
  version: 1.0.0
  contact:
    name: API Support
    email: support@nexus-pm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nexus-pm.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login or register endpoints.
        
        Example: `Authorization: Bearer <your_jwt_token>`

  # ============================================
  # REUSABLE SCHEMAS
  # ============================================
  schemas:
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
      description: User role for authorization
    TaskStatus:
      type: string
      enum:
        - TODO
        - IN_PROGRESS
        - DONE
      description: Current status of a task
    TaskPriority:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
      description: Priority level of a task

    # ========== AUTH SCHEMAS ==========
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique user identifier
        name:
          type: string
          example: John Doe
          description: User's full name
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: User's email address
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp

    # ========== PROJECT SCHEMAS ==========
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique project identifier
        name:
          type: string
          example: Website Redesign
          description: Project name
        description:
          type: string
          nullable: true
          example: Complete redesign of the company website
          description: Project description
        createdById:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: ID of the user who created the project
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Project creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp
        createdBy:
          $ref: '#/components/schemas/User'
        users:
          type: array
          items:
            $ref: '#/components/schemas/ProjectUser'
          description: Users assigned to this project
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: Tasks belonging to this project
    ProjectUser:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
          description: Project ID
        userId:
          type: string
          format: uuid
          description: User ID
        user:
          $ref: '#/components/schemas/User'
...
    CreateTaskRequest:
      type: object
      required:
        - title
        - projectId
        - assigneeId
      properties:
        title:
          type: string
          maxLength: 255
          example: Implement user authentication
          description: Task title
        description:
          type: string
          example: Implement JWT-based authentication system
          description: Detailed task description
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        dueDate:
          type: string
          format: date-time
          example: 2024-02-01T00:00:00Z
          description: Task deadline (ISO 8601 format)
        projectId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: ID of the project this task belongs to
        assigneeId:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
          description: ID of the user assigned to this task
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          example: Implement user authentication with OAuth
          description: Updated task title
        description:
          type: string
          example: Implement JWT and OAuth authentication
          description: Updated task description
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        dueDate:
          type: string
          format: date-time
          example: 2024-02-15T00:00:00Z
          description: Updated task deadline
        assigneeId:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
          description: Updated assignee ID
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique task identifier
        title:
          type: string
          example: Implement user authentication
          description: Task title
        description:
          type: string
          nullable: true
          example: Implement JWT-based authentication system
          description: Task description
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        dueDate:
          type: string
          format: date-time
          nullable: true
          example: 2024-02-01T00:00:00Z
          description: Task deadline
        projectId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Associated project ID
        assigneeId:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
          description: Assigned user ID
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Task creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp
        project:
          $ref: '#/components/schemas/Project'
        assignee:
          $ref: '#/components/schemas/User'
        history:
          type: array
          items:
            $ref: '#/components/schemas/TaskHistory'
          description: Task change history
    TaskHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique history record identifier
        taskId:
          type: string
          format: uuid
          description: Associated task ID
        updatedById:
          type: string
          format: uuid
          description: ID of user who made the change
        oldStatus:
          $ref: '#/components/schemas/TaskStatus'
        newStatus:
          $ref: '#/components/schemas/TaskStatus'
        oldPriority:
          $ref: '#/components/schemas/TaskPriority'
        newPriority:
          $ref: '#/components/schemas/TaskPriority'
        timestamp:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: When the change occurred
        updatedBy:
          $ref: '#/components/schemas/User'

    # ========== USER MANAGEMENT SCHEMAS ==========
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP status code
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: Validation failed
          description: Error message or array of messages
        error:
          type: string
          example: Bad Request
          description: Error type
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        error:
          type: string
          example: Unauthorized
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Forbidden resource
        error:
          type: string
          example: Forbidden
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found
        error:
          type: string
          example: Not Found

# ============================================
# API ENDPOINTS
# ============================================
paths:
...
  /tasks:
    post:
      tags:
        - Tasks
      summary: Create a new task
      description: |
        Create a new task within a project.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can create tasks
        - Task must be assigned to a project and a user
        - Default status is TODO, default priority is MEDIUM
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              basicTask:
                summary: Basic task with required fields
                value:
                  title: Implement user authentication
                  projectId: 550e8400-e29b-41d4-a716-446655440000
                  assigneeId: 660e8400-e29b-41d4-a716-446655440001
              fullTask:
                summary: Task with all fields
                value:
                  title: Design database schema
                  description: Create ERD and implement Prisma schema
                  status: TODO
                  priority: HIGH
                  dueDate: "2024-02-01T00:00:00Z"
                  projectId: 550e8400-e29b-41d4-a716-446655440000
                  assigneeId: 660e8400-e29b-41d4-a716-446655440001
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Project or assignee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Tasks
      summary: Get all tasks
      description: |
        Retrieve all tasks based on user role.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can view all tasks in the system
        - **USER**: Can only view tasks assigned to them
      operationId: getAllTasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
              examples:
                tasks:
                  summary: List of tasks
                  value:
                    - id: 550e8400-e29b-41d4-a716-446655440000
                      title: Implement user authentication
                      description: JWT-based authentication
                      status: IN_PROGRESS
                      priority: HIGH
                      dueDate: "2024-02-01T00:00:00Z"
                      projectId: 770e8400-e29b-41d4-a716-446655440002
                      assigneeId: 660e8400-e29b-41d4-a716-446655440001
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-01-16T14:20:00Z"
                    - id: 660e8400-e29b-41d4-a716-446655440001
                      title: Design database schema
                      description: Create ERD
                      status: TODO
                      priority: MEDIUM
                      dueDate: null
                      projectId: 770e8400-e29b-41d4-a716-446655440002
                      assigneeId: 880e8400-e29b-41d4-a716-446655440003
                      createdAt: "2024-01-15T11:00:00Z"
                      updatedAt: "2024-01-15T11:00:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get a task by ID
      description: |
        Retrieve a single task by its ID.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can view any task
        - **USER**: Can only view tasks assigned to them
      operationId: getTaskById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              examples:
                task:
                  summary: Task details
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    title: Implement user authentication
                    description: JWT-based authentication system
                    status: IN_PROGRESS
                    priority: HIGH
                    dueDate: "2024-02-01T00:00:00Z"
                    projectId: 770e8400-e29b-41d4-a716-446655440002
                    assigneeId: 660e8400-e29b-41d4-a716-446655440001
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-16T14:20:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - Task not assigned to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Tasks
      summary: Update a task
      description: |
        Update an existing task.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can update any task
        - **USER**: Can only update tasks assigned to them
        
        - All fields are optional (partial update)
        - Status and priority changes are tracked in task history
      operationId: updateTask
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            examples:
              updateStatus:
                summary: Update task status
                value:
                  status: IN_PROGRESS
              updatePriority:
                summary: Update task priority
                value:
                  priority: HIGH
              updateMultiple:
                summary: Update multiple fields
                value:
                  title: Implement OAuth authentication
                  description: Add OAuth 2.0 and JWT authentication
                  status: IN_PROGRESS
                  priority: HIGH
                  dueDate: "2024-02-15T00:00:00Z"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - Task not assigned to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: |
        Delete a task by ID.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can delete tasks
        - Cascading delete: All associated task history will be deleted
      operationId: deleteTask
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========== USER MANAGEMENT ENDPOINTS ==========
...